The common package for ec2 is no longer common, remove it
