--- pkg/orchestration/wiring/ec2compute/common/BUILD.bazel
@@ -1,23 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["common.go"],
-    importpath = "github.com/atlassian/voyager/pkg/orchestration/wiring/ec2compute/common",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//:go_default_library",
-        "//pkg/apis/orchestration/v1:go_default_library",
-        "//pkg/execution/plugins/generic/secretplugin:go_default_library",
-        "//pkg/orchestration/wiring/wiringplugin:go_default_library",
-        "//pkg/orchestration/wiring/wiringutil:go_default_library",
-        "//pkg/orchestration/wiring/wiringutil/compute:go_default_library",
-        "//pkg/orchestration/wiring/wiringutil/iam:go_default_library",
-        "//pkg/orchestration/wiring/wiringutil/knownshapes:go_default_library",
-        "//vendor/github.com/atlassian/smith/pkg/apis/smith/v1:go_default_library",
-        "//vendor/github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1:go_default_library",
-        "//vendor/github.com/pkg/errors:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
-    ],
-)
